reworking table structure
