+++ SmartSaver/SmartSaver.csproj
@@ -7,6 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Update="forms\finances\FormExpenses.cs" />
+    <Compile Update="forms\finances\FormIncome.cs" />
     <Compile Update="Properties\Resources.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
