--- SmartSaver/forms/FormExpenses.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Text;
-using System.Windows.Forms;
-
-namespace SmartSaver.forms
-{
-    public partial class FormExpenses : Form
-    {
-        public FormExpenses()
-        {
-            InitializeComponent();
-        }
-    }
-}
