+++ SmartSaver/forms/finances/FormAddIncome.resx
